+++ code_pylint.py
@@ -6,6 +6,7 @@ import math
 from pylint import __version__
 from pylint.lint import Run
 
+
 MIN_NOTE = 9.5
 
 UNWATCHED_ERRORS = ['fixme', 'trailing-whitespace', 'import-error', 'missing-final-newline', 'use-maxsplit-arg']