+++ volmdlr/from_ocp.py
@@ -8,6 +8,7 @@ from OCP.BRep import BRep_Tool
 from OCP.BRepTools import BRepTools, BRepTools_WireExplorer
 from OCP.TopAbs import (TopAbs_EDGE, TopAbs_FACE, TopAbs_VERTEX, TopAbs_WIRE, TopAbs_SHELL, TopAbs_ShapeEnum,
                         TopAbs_SOLID, TopAbs_COMPSOLID, TopAbs_COMPOUND)
+from OCP.ShapeFix import ShapeFix_Shape
 from OCP.TopoDS import (TopoDS_Face, TopoDS_Shell, TopoDS_Wire, TopoDS, TopoDS_Shape)
 
 from OCP.TopExp import TopExp_Explorer
@@ -497,7 +498,7 @@ def get_wires_from_face(face):
     face_wires = get_wires(face)
     if len(face_wires) > 1:
         outer_wire = BRepTools.OuterWire_s(face)
-        inner_wires = [wire for wire in face_wires if not outer_wire.IsSame(wire)]
+        inner_wires = [wire for wire in face_wires if not outer_wire.IsPartner(wire)]
     else:
         outer_wire = face_wires[0]
         inner_wires = []
@@ -516,7 +517,8 @@ def get_contour2d_from_face_wire(contour2d_class, wire, face, occt_to_volmdlr):
         orientation = exp.Current().Orientation()
         if orientation == 1:
             u_start, u_end = u_end, u_start
-
+        if crv.IsInstance("Geom2d_TrimmedCurve"):
+            crv = crv.BasisCurve()
         list_edges.append(volmdlr_edge2d_from_ocp_curve(occt_to_volmdlr[crv.get_type_name_s()], crv,
                                                          u_start, u_end, orientation))
         exp.Next()
@@ -575,3 +577,14 @@ def downcast(obj: TopoDS_Shape) -> TopoDS_Shape:
     f_downcast: Any = downcast_LUT[shapetype(obj)]
 
     return f_downcast(obj)
+
+
+def fix(obj: TopoDS_Shape) -> TopoDS_Shape:
+    """
+    Fix a TopoDS object to suitable specialized type.
+    """
+
+    shape_fixer = ShapeFix_Shape(obj)
+    shape_fixer.Perform()
+
+    return downcast(shape_fixer.Shape())